[FIX] vocabularies not loading on collection screen
