+++ lib/src/common/domain/use_cases/vocabulary/get_vocabularies_to_practise_use_case.dart
@@ -1,14 +1,12 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:vocabualize/src/common/data/repositories/vocabulary_repository_impl.dart';
-import 'package:vocabualize/src/common/domain/entities/filter_options.dart';
 import 'package:vocabualize/src/common/domain/entities/tag.dart';
 import 'package:vocabualize/src/common/domain/entities/vocabulary.dart';
 
 final getVocabulariesToPractiseUseCaseProvider = AutoDisposeProvider.family((ref, Tag? tag) {
-  final filterOptions = tag == null ? null : FilterOptions(tag: tag);
   return GetVocabulariesToPractiseUseCase(
-    vocabularies: ref.watch(vocabularyProvider(filterOptions)).asData?.value ?? [],
-  ).call();
+    vocabularies: ref.watch(vocabularyProvider).value ?? [],
+  ).call(tag);
 });
 
 class GetVocabulariesToPractiseUseCase {
@@ -18,7 +16,10 @@ class GetVocabulariesToPractiseUseCase {
     required List<Vocabulary> vocabularies,
   }) : _vocabularies = vocabularies;
 
-  List<Vocabulary> call() {
-    return _vocabularies.where((vocabulary) => vocabulary.isDue).toList();
+  List<Vocabulary> call(Tag? tag) {
+    return _vocabularies.where((vocabulary) {
+      final containsTag = tag?.id == null || vocabulary.tagIds.contains(tag?.id);
+      return vocabulary.isDue && containsTag;
+    }).toList();
   }
 }
