+++ lib/src/features/collections/presentation/controllers/collection_controller.dart
@@ -19,7 +19,7 @@ class CollectionController extends AutoDisposeFamilyAsyncNotifier<CollectionStat
     final tag = arg;
     return CollectionState(
       tag: tag,
-      tagVocabularies: ref.watch(getVocabulariesUseCaseProvider(FilterOptions(tag: tag))),
+      tagVocabularies: await ref.watch(getVocabulariesUseCaseProvider(FilterOptions(tag: tag))),
       areImagesEnabled: await ref.watch(getAreImagesEnabledUseCaseProvider.future),
     );
   }
